rtorus {alphashape3d}	R Documentation

Generate points in the torus

Description

This function generates n random points within the torus whose minor radius is r, major radius is R and center is ct.

Usage

rtorus(n, r, R, ct = c(0, 0, 0), rotx = NULL)

Arguments

n
Number of observations.

r
Minor radius (radius of the tube).

R
Major radius (distance from the center of the tube to the center of the torus).

ct
A vector with the coordinates of the center of the torus.

rotx
If not NULL, a rotation through an angle rotx (in radians) about the x-axis is performed.

Examples

T1 <- rtorus(2000, 0.5, 2.5)
bbox3d(color = c("white", "black"))
points3d(T1, col = 4)

T2 <- rtorus(2000, 0.5, 2.5, ct = c(2, 0, 0.5), rotx = pi/2)
points3d(T2, col = 2)

[Package alphashape3d version 1.3.3 Index]